requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-chroma>=0.1.0
langchain-huggingface>=0.1.0
chromadb>=0.5.0
pypdf>=4.0.0
streamlit>=1.40.0
google-generativeai>=0.8.0
sentence-transformers>=3.0.0
